tools/outlook-mcp-server/requirements.txt
mcp>=1.0.0
httpx>=0.27.0
pydantic>=2.0.0
requests>=2.31.0
python-dotenv>=1.0.0
